requirements.txt
hyperliquid-python-sdk
ccxt
Pillow
tweepy
